# LaravelProject2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/createLabel": {
            "post": {
                "summary": "create label",
                "description": "create label",
                "operationId": "5fa93c835d46a9760f3ccb556db3e8ac",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "labelname"
                                ],
                                "properties": {
                                    "labelname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Label Added Sucessfully"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/getLableById": {
            "get": {
                "summary": "get Lable By Id",
                "description": "get Label For an Particular User",
                "operationId": "151a5cc2f9284c48b03d4807af92e4e2",
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "All label are found Successfully"
                    },
                    "404": {
                        "description": "Label Not Found"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/updateLabelById": {
            "post": {
                "summary": "update label",
                "description": "update user label",
                "operationId": "caf842567e4f993a6bdf8c55295361ee",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "labelname"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "labelname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Label Updated Successfully"
                    },
                    "404": {
                        "description": "Please enter valid id"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/deleteLabelById": {
            "post": {
                "summary": "Delete Label",
                "description": "Delete User Label",
                "operationId": "d87118f7ed08923cb86a27685358fa36",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label Successfully Deleted"
                    },
                    "404": {
                        "description": "Enter valid id"
                    },
                    "401": {
                        "description": "Invalid Authorization Token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "laravel bearer authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}